+++ opt-env-conf/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed a bug in which unrecognised arguments would be parsed exponentially.
 * The special `--version` command no longer allows any other arguments.
 * Fixed a bug in which some source locations still showed up even though debug mode was not on.
+* Fixed that only one codec for a configuration setting was tried.
 
 ## [0.4.0.1] - 2024-07-26